# schemas/openapi/v1.json
{
  "components": {
    "schemas": {
...
      "ApiErrorEnvelope": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ApiErrorEnvelope",
        "type": "object"
      },
...
      "BatchRequest": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/IngestionRequest"
            },
            "maxItems": 20,
            "minItems": 1,
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "BatchRequest",
        "type": "object"
      },
...
      "CollectionResponse_Source_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/Source"
            },
            "title": "Data",
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          },
          "pagination": {
            "$ref": "#/components/schemas/CursorPagination"
          }
        },
        "required": [
          "data",
          "pagination",
          "meta"
        ],
        "title": "CollectionResponse[Source]",
        "type": "object"
      },
...
      "ErrorDetail": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "message": {
            "minLength": 1,
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
...
      "IngestionRequest": {
        "additionalProperties": false,
        "properties": {
          "content_base64": {
            "maxLength": 349528,
            "minLength": 1,
            "title": "Content Base64",
            "type": "string"
          },
          "content_type": {
            "default": "text/plain",
            "enum": [
              "text/plain",
              "text/html",
              "application/json"
            ],
            "title": "Content Type",
            "type": "string"
          },
          "correlation_id": {
            "default": "local-ingestion",
            "maxLength": 128,
            "minLength": 1,
            "title": "Correlation Id",
            "type": "string"
          },
          "first_seen_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Seen At"
          },
          "idempotency_key": {
            "maxLength": 512,
            "minLength": 1,
            "title": "Idempotency Key",
            "type": "string"
          },
          "language": {
            "anyOf": [
              {
                "maxLength": 35,
                "minLength": 2,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "media": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/MediaInput"
            },
            "maxItems": 4,
            "title": "Media",
            "type": "array"
          },
          "published_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          },
          "source_id": {
            "pattern": "^src_[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$",
            "title": "Source Id",
            "type": "string"
          },
          "source_url": {
            "anyOf": [
              {
                "format": "uri",
                "maxLength": 2083,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Url"
          },
          "title": {
            "anyOf": [
              {
                "maxLength": 512,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "required": [
          "source_id",
          "idempotency_key",
          "content_base64"
        ],
        "title": "IngestionRequest",
        "type": "object"
      },
...
      "MediaInput": {
        "additionalProperties": false,
        "properties": {
          "content_base64": {
            "maxLength": 349528,
            "minLength": 1,
            "title": "Content Base64",
            "type": "string"
          },
          "content_type": {
            "enum": [
              "image/png",
              "image/jpeg",
              "application/pdf",
              "text/plain"
            ],
            "title": "Content Type",
            "type": "string"
          },
          "kind": {
            "enum": [
              "image",
              "attachment"
            ],
            "title": "Kind",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "content_type",
          "content_base64"
        ],
        "title": "MediaInput",
        "type": "object"
      },
      "NewsDocument": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "document_id": {
            "pattern": "^doc_[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$",
            "title": "Document Id",
            "type": "string"
          },
          "first_seen_at": {
            "format": "date-time",
            "title": "First Seen At",
            "type": "string"
          },
          "ingested_at": {
            "format": "date-time",
            "title": "Ingested At",
            "type": "string"
          },
          "ingestion_id": {
            "pattern": "^ing_[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$",
            "title": "Ingestion Id",
            "type": "string"
          },
          "language": {
            "anyOf": [
              {
                "maxLength": 35,
                "minLength": 2,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "published_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          },
          "revision": {
            "default": 1,
            "minimum": 1.0,
            "title": "Revision",
            "type": "integer"
          },
          "schema_version": {
            "const": "1",
            "default": "1",
            "title": "Schema Version",
            "type": "string"
          },
          "source_id": {
            "pattern": "^src_[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$",
            "title": "Source Id",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          },
          "title": {
            "maxLength": 512,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "document_id",
          "ingestion_id",
          "source_id",
          "title",
          "text",
          "first_seen_at",
          "ingested_at",
          "created_at"
        ],
        "title": "NewsDocument",
        "type": "object"
      },
...
      "RawIngestion": {
        "additionalProperties": false,
        "properties": {
          "first_seen_at": {
            "format": "date-time",
            "title": "First Seen At",
            "type": "string"
          },
          "idempotency_key": {
            "maxLength": 512,
            "minLength": 1,
            "title": "Idempotency Key",
            "type": "string"
          },
          "ingested_at": {
            "format": "date-time",
            "title": "Ingested At",
            "type": "string"
          },
          "ingestion_id": {
            "pattern": "^ing_[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$",
            "title": "Ingestion Id",
            "type": "string"
          },
          "object": {
            "$ref": "#/components/schemas/ObjectReference"
          },
          "published_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          },
          "raw_object_id": {
            "pattern": "^raw_[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$",
            "title": "Raw Object Id",
            "type": "string"
          },
          "schema_version": {
            "const": "1",
            "default": "1",
            "title": "Schema Version",
            "type": "string"
          },
          "source_id": {
            "pattern": "^src_[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$",
            "title": "Source Id",
            "type": "string"
          },
          "source_url": {
            "anyOf": [
              {
                "format": "uri",
                "maxLength": 2083,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Url"
          }
        },
        "required": [
          "ingestion_id",
          "source_id",
          "raw_object_id",
          "object",
          "first_seen_at",
          "ingested_at",
          "idempotency_key"
        ],
        "title": "RawIngestion",
        "type": "object"
      },
...
      "ResponseMeta": {
        "additionalProperties": false,
        "properties": {
          "api_version": {
            "const": "v1",
            "default": "v1",
            "title": "Api Version",
            "type": "string"
          },
          "request_id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "request_id"
        ],
        "title": "ResponseMeta",
        "type": "object"
      },
...
      "SingleResponse_NewsDocument_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/NewsDocument"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "SingleResponse[NewsDocument]",
        "type": "object"
      },
...
      "SingleResponse_RawIngestion_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/RawIngestion"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "SingleResponse[RawIngestion]",
        "type": "object"
      },
...
      "SingleResponse_Source_": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Source"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "SingleResponse[Source]",
        "type": "object"
      },
...
      "SingleResponse_dict_str__Any__": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "additionalProperties": true,
            "title": "Data",
            "type": "object"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "SingleResponse[dict[str, Any]]",
        "type": "object"
      },
      "SingleResponse_dict_str__str__": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Data",
            "type": "object"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "SingleResponse[dict[str, str]]",
        "type": "object"
      },
      "Source": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "kind": {
            "enum": [
              "feed",
              "api",
              "upload",
              "web"
            ],
            "title": "Kind",
            "type": "string"
          },
          "name": {
            "maxLength": 512,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "schema_version": {
            "const": "1",
            "default": "1",
            "title": "Schema Version",
            "type": "string"
          },
          "source_id": {
            "pattern": "^src_[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$",
            "title": "Source Id",
            "type": "string"
          },
          "url": {
            "anyOf": [
              {
                "format": "uri",
                "maxLength": 2083,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          }
        },
        "required": [
          "source_id",
          "name",
          "kind",
          "created_at"
        ],
        "title": "Source",
        "type": "object"
      },
      "SourceCreate": {
        "additionalProperties": false,
        "properties": {
          "kind": {
            "enum": [
              "feed",
              "api",
              "upload",
              "web"
            ],
            "title": "Kind",
            "type": "string"
          },
          "name": {
            "maxLength": 512,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "url": {
            "anyOf": [
              {
                "format": "uri",
                "maxLength": 2083,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          }
        },
        "required": [
          "name",
          "kind"
        ],
        "title": "SourceCreate",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Modular monolith with durable offline ingestion and canonical document retrieval.",
    "title": "AegisNews",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/documents/{document_id}": {
      "get": {
        "operationId": "get_document_api_v1_documents__document_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "title": "Document Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "as_of",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "As Of"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SingleResponse_NewsDocument_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Document",
        "tags": [
          "ingestion"
        ]
      }
    },
...
    "/api/v1/ingestion-runs/{workflow_id}": {
      "get": {
        "operationId": "get_run_api_v1_ingestion_runs__workflow_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "workflow_id",
            "required": true,
            "schema": {
              "title": "Workflow Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SingleResponse_dict_str__Any__"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Run",
        "tags": [
          "ingestion"
        ]
      }
    },
    "/api/v1/ingestions": {
      "post": {
        "operationId": "submit_api_v1_ingestions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngestionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SingleResponse_dict_str__str__"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Submit",
        "tags": [
          "ingestion"
        ]
      }
    },
    "/api/v1/ingestions/batch": {
      "post": {
        "operationId": "submit_batch_api_v1_ingestions_batch_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SingleResponse_dict_str__Any__"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Submit Batch",
        "tags": [
          "ingestion"
        ]
      }
    },
    "/api/v1/ingestions/{ingestion_id}": {
      "get": {
        "operationId": "get_ingestion_api_v1_ingestions__ingestion_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "ingestion_id",
            "required": true,
            "schema": {
              "title": "Ingestion Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SingleResponse_RawIngestion_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Ingestion",
        "tags": [
          "ingestion"
        ]
      }
    },
...
    "/api/v1/sources": {
      "get": {
        "operationId": "sources_api_v1_sources_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 128,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionResponse_Source_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Sources",
        "tags": [
          "product"
        ]
      },
      "post": {
        "operationId": "create_source_api_v1_sources_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SourceCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SingleResponse_Source_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Source",
        "tags": [
          "ingestion"
        ]
      }
    },
    "/api/v1/sources/{source_id}": {
      "get": {
        "operationId": "get_source_api_v1_sources__source_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "source_id",
            "required": true,
            "schema": {
              "title": "Source Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SingleResponse_Source_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Source",
        "tags": [
          "ingestion"
        ]
      }
    },
...
  },
...
}